backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
email-validator
dnspython
